requirements.txt
flask
pydantic
mysql-connector